+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Added 3rd argument to `highlight(code, lang, attrs)`, #626.
 - Rewrite tables according to latest GFM spec, #697.
+- Use `rollup.js` to browserify sources.
 
 ### Fixed
 - Fix mappings for table rows (amended fix made in 11.0.1), #705.